requirements.txt
numpy>=1.20.0
scipy>=1.7.0
scikit-learn>=1.0.0
faiss-cpu>=1.7.0
tqdm>=4.60.0

# Optional LLM support
openai>=1.0.0
anthropic>=0.5.0

# Development
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
